Setup Redux Store

+++ app/api/review_routes.py
@@ -82,8 +82,13 @@ def edit(id):
 @login_required
 def delete(id):
     review = Review.query.get(id)
-    db.session.delete(review)
-    db.commit()
+    if (current_user.id == review.user_id) {
+        db.session.delete(review)
+        db.commit()
+    } else {
+        return {"error": "You are not authorized to delete this review"}
+    }
+
     return review.to_dict()
 
 @review_routes.route('/unauthorized')
